# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Kokpit Okulları API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface LoginBody {
  password: string;
}

export interface LoginResponse {
  ok: boolean;
}

export interface MeResponse {
  loggedIn: boolean;
}

export interface Haber {
  id: number;
  baslik: string;
  ozet: string;
  icerik: string;
  /** @nullable */
  resimUrl?: string | null;
  kategori: string;
  yayinda: boolean;
  onemli: boolean;
  yayinTarihi: string;
  createdAt: string;
  updatedAt: string;
}

export interface CreateHaberBody {
  baslik: string;
  ozet: string;
  icerik: string;
  /** @nullable */
  resimUrl?: string | null;
  kategori: string;
  yayinda?: boolean;
  onemli?: boolean;
  yayinTarihi?: string;
}

export interface UpdateHaberBody {
  baslik?: string;
  ozet?: string;
  icerik?: string;
  /** @nullable */
  resimUrl?: string | null;
  kategori?: string;
  yayinda?: boolean;
  onemli?: boolean;
  yayinTarihi?: string;
}

export interface Duyuru {
  id: number;
  baslik: string;
  icerik: string;
  onemli: boolean;
  yayinda: boolean;
  yayinTarihi: string;
  createdAt: string;
  updatedAt: string;
}

export interface CreateDuyuruBody {
  baslik: string;
  icerik: string;
  onemli?: boolean;
  yayinda?: boolean;
  yayinTarihi?: string;
}

export interface UpdateDuyuruBody {
  baslik?: string;
  icerik?: string;
  onemli?: boolean;
  yayinda?: boolean;
  yayinTarihi?: string;
}

export interface Etkinlik {
  id: number;
  baslik: string;
  aciklama: string;
  /** @nullable */
  resimUrl?: string | null;
  tarih: string;
  /** @nullable */
  yer?: string | null;
  kategori: string;
  yayinda: boolean;
  createdAt: string;
  updatedAt: string;
}

export interface CreateEtkinlikBody {
  baslik: string;
  aciklama: string;
  /** @nullable */
  resimUrl?: string | null;
  tarih: string;
  /** @nullable */
  yer?: string | null;
  kategori: string;
  yayinda?: boolean;
}

export interface UpdateEtkinlikBody {
  baslik?: string;
  aciklama?: string;
  /** @nullable */
  resimUrl?: string | null;
  tarih?: string;
  /** @nullable */
  yer?: string | null;
  kategori?: string;
  yayinda?: boolean;
}

export interface FotoGaleri {
  id: number;
  baslik: string;
  resimUrl: string;
  kategori: string;
  yayinda: boolean;
  sira: number;
  createdAt: string;
  updatedAt: string;
}

export interface CreateFotoGaleriBody {
  baslik: string;
  resimUrl: string;
  kategori: string;
  yayinda?: boolean;
  sira?: number;
}

export interface UpdateFotoGaleriBody {
  baslik?: string;
  resimUrl?: string;
  kategori?: string;
  yayinda?: boolean;
  sira?: number;
}

export interface VideoGaleri {
  id: number;
  baslik: string;
  youtubeId: string;
  kategori: string;
  yayinda: boolean;
  sira: number;
  createdAt: string;
  updatedAt: string;
}

export interface CreateVideoGaleriBody {
  baslik: string;
  youtubeId: string;
  kategori: string;
  yayinda?: boolean;
  sira?: number;
}

export interface UpdateVideoGaleriBody {
  baslik?: string;
  youtubeId?: string;
  kategori?: string;
  yayinda?: boolean;
  sira?: number;
}

export type ListHaberlerParams = {
  kategori?: string;
  yayinda?: string;
};

export type ListDuyurularParams = {
  yayinda?: string;
};

export type ListEtkinliklerParams = {
  yayinda?: string;
};

export type ListFotoGaleriParams = {
  kategori?: string;
  yayinda?: string;
};

export type ListVideoGaleriParams = {
  kategori?: string;
  yayinda?: string;
};