# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Kokpit Okulları API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: adminLogin
      tags: [auth]
      summary: Admin girişi
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Giriş başarılı
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          description: Geçersiz şifre
...
  /duyurular:
    get:
      operationId: listDuyurular
      tags: [duyurular]
      summary: Tüm duyurular
      parameters:
        - name: yayinda
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Duyuru listesi
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Duyuru"
    post:
      operationId: createDuyuru
      tags: [duyurular]
      summary: Yeni duyuru oluştur
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDuyuruBody"
      responses:
        "201":
          description: Oluşturuldu
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Duyuru"
  /duyurular/{id}:
    get:
      operationId: getDuyuru
      tags: [duyurular]
      summary: Duyuru detayı
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Duyuru
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Duyuru"
        "404":
          description: Bulunamadı
    patch:
      operationId: updateDuyuru
      tags: [duyurular]
      summary: Duyuruyu güncelle
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateDuyuruBody"
      responses:
        "200":
          description: Güncellendi
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Duyuru"
        "404":
          description: Bulunamadı
    delete:
      operationId: deleteDuyuru
      tags: [duyurular]
      summary: Duyuruyu sil
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Silindi
        "404":
          description: Bulunamadı
  /etkinlikler:
    get:
      operationId: listEtkinlikler
      tags: [etkinlikler]
      summary: Tüm etkinlikler
      parameters:
        - name: yayinda
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Etkinlik listesi
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Etkinlik"
    post:
      operationId: createEtkinlik
      tags: [etkinlikler]
      summary: Yeni etkinlik oluştur
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEtkinlikBody"
      responses:
        "201":
          description: Oluşturuldu
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Etkinlik"
  /etkinlikler/{id}:
    get:
      operationId: getEtkinlik
      tags: [etkinlikler]
      summary: Etkinlik detayı
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Etkinlik
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Etkinlik"
        "404":
          description: Bulunamadı
    patch:
      operationId: updateEtkinlik
      tags: [etkinlikler]
      summary: Etkinliği güncelle
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateEtkinlikBody"
      responses:
        "200":
          description: Güncellendi
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Etkinlik"
        "404":
          description: Bulunamadı
    delete:
      operationId: deleteEtkinlik
      tags: [etkinlikler]
      summary: Etkinliği sil
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Silindi
        "404":
          description: Bulunamadı
...
components:
  schemas:
...
    LoginBody:
      type: object
      properties:
        password:
          type: string
      required:
        - password
    LoginResponse:
      type: object
      properties:
        ok:
          type: boolean
      required:
        - ok
...
    Duyuru:
      type: object
      properties:
        id:
          type: integer
        baslik:
          type: string
        icerik:
          type: string
        onemli:
          type: boolean
        yayinda:
          type: boolean
        yayinTarihi:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - baslik
        - icerik
        - onemli
        - yayinda
        - yayinTarihi
        - createdAt
        - updatedAt
    CreateDuyuruBody:
      type: object
      properties:
        baslik:
          type: string
        icerik:
          type: string
        onemli:
          type: boolean
        yayinda:
          type: boolean
        yayinTarihi:
          type: string
      required:
        - baslik
        - icerik
    UpdateDuyuruBody:
      type: object
      properties:
        baslik:
          type: string
        icerik:
          type: string
        onemli:
          type: boolean
        yayinda:
          type: boolean
        yayinTarihi:
          type: string
    Etkinlik:
      type: object
      properties:
        id:
          type: integer
        baslik:
          type: string
        aciklama:
          type: string
        resimUrl:
          type: ["string", "null"]
        tarih:
          type: string
        yer:
          type: ["string", "null"]
        kategori:
          type: string
        yayinda:
          type: boolean
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - baslik
        - aciklama
        - tarih
        - kategori
        - yayinda
        - createdAt
        - updatedAt
    CreateEtkinlikBody:
      type: object
      properties:
        baslik:
          type: string
        aciklama:
          type: string
        resimUrl:
          type: ["string", "null"]
        tarih:
          type: string
        yer:
          type: ["string", "null"]
        kategori:
          type: string
        yayinda:
          type: boolean
      required:
        - baslik
        - aciklama
        - tarih
        - kategori
    UpdateEtkinlikBody:
      type: object
      properties:
        baslik:
          type: string
        aciklama:
          type: string
        resimUrl:
          type: ["string", "null"]
        tarih:
          type: string
        yer:
          type: ["string", "null"]
        kategori:
          type: string
        yayinda:
          type: boolean
...